+++ aptos-move/block-executor/src/lib.rs
@@ -2,8 +2,6 @@
 // Parts of the project are originally copyright © Meta Platforms, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-use aptos_mvhashmap::types::TxnIndex;
-
 /**
 The high level parallel execution logic is implemented in 'executor.rs'. The
 input of parallel executor is a block of transactions, containing a sequence
@@ -141,6 +139,7 @@ subsequent incarnation to finish.
 pub mod counters;
 pub mod errors;
 pub mod executor;
+pub mod index_mapping;
 #[cfg(any(test, feature = "fuzzing"))]
 pub mod proptest_types;
 mod scheduler;
@@ -149,94 +148,3 @@ mod txn_last_input_output;
 #[cfg(test)]
 mod unit_tests;
 pub mod view;
-
-/// In standard BlockSTM, the internal states assume contiguous transactions indices.
-/// To also support sharded execution (where each shard gets a subsequence of the transactions
-/// and possibly waits for transaction results from other shards),
-/// an index mapping is need as an additional input to our existing BlockSTM implementation.
-#[derive(Clone)]
-pub struct IndexMapping {
-    indices: Vec<usize>,
-    /// A TxnIndex -> local position mapping.
-    /// Currently implemented as a `Vec` of size equal to the block size, assuming it's not too large.
-    positions_by_index: Vec<usize>,
-    /// A fake index representing the end of the transaction list.
-    end_index: usize,
-}
-
-impl IndexMapping {
-    /// Create Positions by TxnIndex mapping from the TxnIndex list.
-    pub fn inverse(block_size: usize, indices: &[usize]) -> Vec<usize> {
-        let mut ret = vec![usize::MAX; block_size];
-        for (pos, &index) in indices.iter().enumerate() {
-            ret[index] = pos;
-        }
-        ret
-    }
-
-    pub fn new(indices: Vec<usize>, block_size: usize) -> Self {
-        let positions_by_index = Self::inverse(block_size, &indices);
-        Self {
-            indices,
-            positions_by_index,
-            end_index: block_size,
-        }
-    }
-
-    pub fn wrap(indices: Vec<usize>, positions_by_idx: Vec<usize>) -> Self {
-        let end_index = positions_by_idx.len();
-        Self {
-            indices,
-            positions_by_index: positions_by_idx,
-            end_index,
-        }
-    }
-
-    pub fn new_unsharded(block_size: usize) -> Self {
-        Self {
-            indices: (0..block_size).collect(),
-            positions_by_index: (0..block_size).collect(),
-            end_index: block_size,
-        }
-    }
-
-    pub fn next_index(&self, index: TxnIndex) -> TxnIndex {
-        let index = index as usize;
-        let ret = if index == self.end_index {
-            self.end_index
-        } else {
-            let pos = self.positions_by_index[index];
-            if pos >= self.indices.len() - 1 {
-                self.end_index
-            } else {
-                self.indices[pos + 1]
-            }
-        };
-        ret as TxnIndex
-    }
-
-    pub fn num_txns(&self) -> usize {
-        self.indices.len()
-    }
-
-    pub fn iter_txn_indices(&self) -> Box<dyn Iterator<Item = TxnIndex> + '_> {
-        Box::new(self.indices.iter().map(|&i| i as TxnIndex))
-    }
-
-    pub fn position_by_index(&self, index: TxnIndex) -> Option<usize> {
-        let index = index as usize;
-        if index >= self.positions_by_index.len() {
-            None
-        } else {
-            Some(self.positions_by_index[index])
-        }
-    }
-
-    pub fn end_index(&self) -> TxnIndex {
-        self.end_index as TxnIndex
-    }
-
-    pub fn index(&self, i: usize) -> TxnIndex {
-        self.indices[i] as TxnIndex
-    }
-}
