+++ aptos-move/block-executor/src/proptest_types/bencher.rs
@@ -129,11 +129,7 @@ where
             EmptyDataView<KeyType<K>, ValueType<V>>,
             ExecutableTestType,
         >::new(num_cpus::get(), executor_thread_pool, None)
-        .execute_transactions_parallel(
-            (),
-            &self.transactions,
-            &data_view,
-        );
+        .execute_transactions_parallel((), &self.transactions, &data_view);
 
         self.expected_output.assert_output(&output);
     }
