+++ aptos-move/block-executor/src/proptest_types/tests.rs
@@ -74,11 +74,7 @@ fn run_transactions<K, V>(
             executor_thread_pool.clone(),
             maybe_block_gas_limit,
         )
-        .execute_transactions_parallel(
-            (),
-            &executable_txns,
-            &data_view,
-        );
+        .execute_transactions_parallel((), &executable_txns, &data_view);
 
         if module_access.0 && module_access.1 {
             assert_eq!(output.unwrap_err(), Error::ModulePathReadWrite);
@@ -218,11 +214,7 @@ fn deltas_writes_mixed_with_block_gas_limit(num_txns: usize, maybe_block_gas_lim
             executor_thread_pool.clone(),
             maybe_block_gas_limit,
         )
-        .execute_transactions_parallel(
-            (),
-            &executable_txns,
-            &data_view,
-        );
+        .execute_transactions_parallel((), &executable_txns, &data_view);
 
         let baseline = ExpectedOutput::generate_baseline(
             executable_txns.get_unsharded_transactions().unwrap(),
@@ -278,11 +270,7 @@ fn deltas_resolver_with_block_gas_limit(num_txns: usize, maybe_block_gas_limit: 
             executor_thread_pool.clone(),
             maybe_block_gas_limit,
         )
-        .execute_transactions_parallel(
-            (),
-            &executable_txns,
-            &data_view,
-        );
+        .execute_transactions_parallel((), &executable_txns, &data_view);
 
         let delta_writes = output
             .as_ref()
@@ -454,11 +442,7 @@ fn publishing_fixed_params_with_block_gas_limit(
         DeltaDataView<KeyType<[u8; 32]>, ValueType<[u8; 32]>>,
         ExecutableTestType,
     >::new(num_cpus::get(), executor_thread_pool, maybe_block_gas_limit)
-    .execute_transactions_parallel(
-        (),
-        &executable_txns,
-        &data_view,
-    );
+    .execute_transactions_parallel((), &executable_txns, &data_view);
     assert_ok!(output);
 
     // Adjust the reads of txn indices[2] to contain module read to key 42.
@@ -508,11 +492,7 @@ fn publishing_fixed_params_with_block_gas_limit(
             executor_thread_pool.clone(),
             Some(max(w_index, r_index) as u64 + 1),
         ) // Ensure enough gas limit to commit the module txns
-        .execute_transactions_parallel(
-            (),
-            &executable_txns,
-            &data_view,
-        );
+        .execute_transactions_parallel((), &executable_txns, &data_view);
 
         assert_eq!(output.unwrap_err(), Error::ModulePathReadWrite);
     }
